[BE/FEAT] 요가 수행 시, /sequence, /pose, /start API 생성 (#7)

+++ backend/src/main/java/com/prana/backend/sequence/repository/SequenceRepository.java
@@ -1,8 +1,8 @@
 package com.prana.backend.sequence.repository;
 
-import com.prana.backend.sequence.Sequence;
+import com.prana.backend.entity.Sequence;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SequenceRepository extends JpaRepository<Sequence, Long> {
+public interface SequenceRepository extends JpaRepository<Sequence, Integer> {
 
 }
